Workflow pipeline targets

Target definitions for workflow input.

targetsWorkflow(x)

# S4 method for Workflow
targetsWorkflow(x)

targetsInput(x)

# S4 method for FilePathInput
targetsInput(x)

# S4 method for GroverInput
targetsInput(x)

# S4 method for Workflow
targetsInput(x)

targetsSpectralProcessing(x)

targetsPretreatment(x)

targetsMFassignment(x)

targetsModelling(x)

targetsCorrelations(x)

targetsReport(x)

Arguments

x
the workflow type or an S4 object of class Workflow, FilePathInput or GroverInput

Value

A list of Target S4 class target definitions.

Examples

## Full workflow example
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- inputFilePath(file_paths,sample_information)

workflow_definition <- defineWorkflow(workflow_input,
                                     'FIE-HRMS fingerprinting',
                                     'Example project')

targetsWorkflow(workflow_definition)

#> $input
#> $input$file_paths_list
#> ## Retrieve data file paths
#> tar_file(
#>   file_paths_list,
#>   "data/file_paths.txt"
#> )
#> 
#> $input$mzML
#> ## Track individual data files
#> tar_files(
#>   mzML,
#>   readLines(file_paths_list)
#> )
#> 
#> $input$sample_information_file
#> ## Sample information file path
#> tar_file(
#>   sample_information_file,
#>   "data/runinfo.csv"
#> )
#> 
#> $input$sample_information
#> ## Parse sample information
#> tar_target(
#>   sample_information,
#>   readr::read_csv(sample_information_file)
#> )
#> 
#> 
#> $spectral_processing
#> $spectral_processing$spectral_processing_parameters
#> ## Detect spectral binning parameters
#> tar_target(
#>   spectral_processing_parameters,
#>   binneR::detectParameters(mzML)
#> )
#> 
#> $spectral_processing$spectral_processed
#> ## Perform spectral binning
#> tar_target(
#>   spectral_processed,
#>   binneR::binneRlyse(mzML, sample_information, spectral_processing_parameters),
#>   memory = "transient"
#> )
#> 
#> $spectral_processing$plot_fingerprint
#> ## Plot average spectrum fingerprint
#> tar_target(
#>   plot_fingerprint,
#>   binneR::plotFingerprint(spectral_processed)
#> )
#> 
#> $spectral_processing$plot_chromatogram
#> ## Plot average infusion chromatogram
#> tar_target(
#>   plot_chromatogram,
#>   binneR::plotChromatogram(spectral_processed)
#> )
#> 
#> $spectral_processing$plot_TIC
#> ## Plot sample total ion counts by randomised block
#> tar_target(
#>   plot_TIC,
#>   binneR::plotTIC(spectral_processed)
#> )
#> 
#> $spectral_processing$plot_purity_dist
#> ## Plot bin purity distribution
#> tar_target(
#>   plot_purity_dist,
#>   binneR::plotPurity(spectral_processed)
#> )
#> 
#> $spectral_processing$plot_centrality_dist
#> ## Plot bin centrality distribution
#> tar_target(
#>   plot_centrality_dist,
#>   binneR::plotCentrality(spectral_processed)
#> )
#> 
#> $spectral_processing$summary_processed_features
#> ## Summary of spectrally binned features
#> tar_target(
#>   summary_processed_features,
#>   metaboMisc::featureSummary(spectral_processed)
#> )
#> 
#> $spectral_processing$export_processed_data
#> ## Export spectrally binned data
#> tar_files(
#>   export_processed_data,
#>   metaboMisc::export(spectral_processed, outPath = "exports/spectral_processing")
#> )
#> 
#> 
#> $pre_treatment
#> $pre_treatment$pre_treatment_parameters
#> ## Detect pre-treatment routine parameters
#> tar_target(
#>   pre_treatment_parameters,
#>   metaboMisc::detectPretreatmentParameters(spectral_processed)
#> )
#> 
#> $pre_treatment$pre_treated
#> ## Perform data pre-treatment
#> tar_target(
#>   pre_treated,
#>   metaboMisc::preTreatModes(spectral_processed, pre_treatment_parameters)
#> )
#> 
#> $pre_treatment$export_pre_treated_data
#> ## Export pre-treated data
#> tar_file(
#>   export_pre_treated,
#>   metaboMisc::exportData(pre_treated, type = "pre-treated", outPath = "exports/pre-treated")
#> )
#> 
#> $pre_treatment$export_pre_treated_sample_info
#> ## Export sample information of pre-treated data
#> tar_file(
#>   export_pre_treated_sample_info,
#>   metaboMisc::exportSampleInfo(pre_treated, outPath = "exports/pre-treated")
#> )
#> 
#> $pre_treatment$plot_PCA
#> ## Plot Principle Component Analysis
#> tar_target(
#>   plot_PCA,
#>   metabolyseR::plotPCA(pre_treated, type = "pre-treated")
#> )
#> 
#> $pre_treatment$plot_LDA
#> ## Plot Priniciple Component Analysis - Linear Discriminant Analysis
#> tar_target(
#>   plot_LDA,
#>   metabolyseR::plotLDA(pre_treated, type = "pre-treated")
#> )
#> 
#> $pre_treatment$plot_unsupervised_RF
#> ## Plot multidimensional scaling plot of unsupervised random forest
#> tar_target(
#>   plot_unsupervised_RF,
#>   metabolyseR::plotUnsupervisedRF(pre_treated, type = "pre-treated")
#> )
#> 
#> $pre_treatment$plot_supervised_RF
#> ## Plot multidimensional scaling plot of supervised random forest
#> tar_target(
#>   plot_supervised_RF,
#>   metabolyseR::plotSupervisedRF(pre_treated, type = "pre-treated")
#> )
#> 
#> 
#> $molecular_formula_assignment
#> $molecular_formula_assignment$molecular_formula_assignment_parameters
#> ## Generate molecular formula assignment parameters
#> tar_target(
#>   molecular_formula_assignment_parameters,
#>   MFassign::assignmentParameters("FIE")
#> )
#> 
#> $molecular_formula_assignment$molecular_formula_assignment
#> ## Perform molecular formula assignment
#> tar_target(
#>   molecular_formula_assignment,
#>   pre_treated %>% metabolyseR::dat(type = "pre-treated") %>% MFassign::assignMFs(molecular_formula_assignment_parameters),
#>   memory = "transient"
#> )
#> 
#> $molecular_formula_assignment$assigned_data
#> ## Retieve pre-treated data with molecular formula assignments added to the feature names
#> tar_target(
#>   assigned_data,
#>   metaboMisc::addAssignments(pre_treated, molecular_formula_assignment)
#> )
#> 
#> $molecular_formula_assignment$summary_assignments
#> ## Summarise the assigned molecular formulas
#> tar_target(
#>   summary_assignments,
#>   MFassign::summariseAssignment(molecular_formula_assignment)
#> )
#> 
#> $molecular_formula_assignment$export_assignments
#> ## Export molecular formula assignments
#> tar_files(
#>   export_assignments,
#>   metaboMisc::export(molecular_formula_assignment, outPath = "exports/molecular_formula_assignments")
#> )
#> 
#> 
#> $modelling
#> $modelling$modelling_parameters
#> ## Detect appropriate modelling parameters
#> tar_target(
#>   modelling_parameters,
#>   metaboMisc::detectModellingParameters(assigned_data, cls = "class")
#> )
#> 
#> $modelling$modelling
#> ## Perform modelling
#> tar_target(
#>   modelling,
#>   metabolyseR::reAnalyse(assigned_data, modelling_parameters)
#> )
#> 
#> $modelling$plot_explanatory_heatmap
#> ## Plot a heat map of explanatory features
#> tar_target(
#>   plot_explanatory_heatmap,
#>   metabolyseR::plotExplanatoryHeatmap(modelling)
#> )
#> 
#> $modelling$summary_modelling_metrics
#> ## Retrieve modelling metrics
#> tar_target(
#>   summary_model_metrics,
#>   metabolyseR::metrics(modelling)
#> )
#> 
#> $modelling$summary_modelling_importance
#> ## Retireve modelling feature importance
#> tar_target(
#>   summary_model_importance,
#>   metabolyseR::importance(modelling)
#> )
#> 
#> $modelling$export_modelling
#> ## Export modelling results
#> tar_files(
#>   export_modelling,
#>   metaboMisc::exportModelling(modelling, outPath = "exports/modelling")
#> )
#> 
#> 
#> $correlations
#> $correlations$correlations_parameters
#> ## Generate parameters for correlation analysis
#> tar_target(
#>   correlations_parameters,
#>   metabolyseR::analysisParameters("correlations")
#> )
#> 
#> $correlations$correlations
#> ## Perform correlation analysis
#> tar_target(
#>   correlations,
#>   metabolyseR::reAnalyse(assigned_data, correlations_parameters)
#> )
#> 
#> $correlations$summary_correlations
#> ## Retrieve correlation analysis results
#> tar_target(
#>   summary_correlations,
#>   metabolyseR::analysisResults(correlations, "correlations")
#> )
#> 
#> $correlations$export_correlations
#> ## Export correlation analysis results
#> tar_files(
#>   export_correlations,
#>   metaboMisc::exportCorrelations(correlations, outPath = "exports/correlations")
#> )
#> 
#> 
#> $report
#> $report$report
#> tar_render(
#>   report,
#>   "report/report.Rmd",
#>   output_dir = "exports"
#> )
#> 
#> 

## Examples for individual modules
targetsSpectralProcessing('FIE-HRMS fingerprinting')

#> $spectral_processing_parameters
#> ## Detect spectral binning parameters
#> tar_target(
#>   spectral_processing_parameters,
#>   binneR::detectParameters(mzML)
#> )
#> 
#> $spectral_processed
#> ## Perform spectral binning
#> tar_target(
#>   spectral_processed,
#>   binneR::binneRlyse(mzML, sample_information, spectral_processing_parameters),
#>   memory = "transient"
#> )
#> 
#> $plot_fingerprint
#> ## Plot average spectrum fingerprint
#> tar_target(
#>   plot_fingerprint,
#>   binneR::plotFingerprint(spectral_processed)
#> )
#> 
#> $plot_chromatogram
#> ## Plot average infusion chromatogram
#> tar_target(
#>   plot_chromatogram,
#>   binneR::plotChromatogram(spectral_processed)
#> )
#> 
#> $plot_TIC
#> ## Plot sample total ion counts by randomised block
#> tar_target(
#>   plot_TIC,
#>   binneR::plotTIC(spectral_processed)
#> )
#> 
#> $plot_purity_dist
#> ## Plot bin purity distribution
#> tar_target(
#>   plot_purity_dist,
#>   binneR::plotPurity(spectral_processed)
#> )
#> 
#> $plot_centrality_dist
#> ## Plot bin centrality distribution
#> tar_target(
#>   plot_centrality_dist,
#>   binneR::plotCentrality(spectral_processed)
#> )
#> 
#> $summary_processed_features
#> ## Summary of spectrally binned features
#> tar_target(
#>   summary_processed_features,
#>   metaboMisc::featureSummary(spectral_processed)
#> )
#> 
#> $export_processed_data
#> ## Export spectrally binned data
#> tar_files(
#>   export_processed_data,
#>   metaboMisc::export(spectral_processed, outPath = "exports/spectral_processing")
#> )
#> 

targetsInput(workflow_input)

#> $file_paths_list
#> ## Retrieve data file paths
#> tar_file(
#>   file_paths_list,
#>   "data/file_paths.txt"
#> )
#> 
#> $mzML
#> ## Track individual data files
#> tar_files(
#>   mzML,
#>   readLines(file_paths_list)
#> )
#> 
#> $sample_information_file
#> ## Sample information file path
#> tar_file(
#>   sample_information_file,
#>   "data/runinfo.csv"
#> )
#> 
#> $sample_information
#> ## Parse sample information
#> tar_target(
#>   sample_information,
#>   readr::read_csv(sample_information_file)
#> )
#> 

targetsPretreatment('FIE-HRMS fingerprinting')

#> $pre_treatment_parameters
#> ## Detect pre-treatment routine parameters
#> tar_target(
#>   pre_treatment_parameters,
#>   metaboMisc::detectPretreatmentParameters(spectral_processed)
#> )
#> 
#> $pre_treated
#> ## Perform data pre-treatment
#> tar_target(
#>   pre_treated,
#>   metaboMisc::preTreatModes(spectral_processed, pre_treatment_parameters)
#> )
#> 
#> $export_pre_treated_data
#> ## Export pre-treated data
#> tar_file(
#>   export_pre_treated,
#>   metaboMisc::exportData(pre_treated, type = "pre-treated", outPath = "exports/pre-treated")
#> )
#> 
#> $export_pre_treated_sample_info
#> ## Export sample information of pre-treated data
#> tar_file(
#>   export_pre_treated_sample_info,
#>   metaboMisc::exportSampleInfo(pre_treated, outPath = "exports/pre-treated")
#> )
#> 
#> $plot_PCA
#> ## Plot Principle Component Analysis
#> tar_target(
#>   plot_PCA,
#>   metabolyseR::plotPCA(pre_treated, type = "pre-treated")
#> )
#> 
#> $plot_LDA
#> ## Plot Priniciple Component Analysis - Linear Discriminant Analysis
#> tar_target(
#>   plot_LDA,
#>   metabolyseR::plotLDA(pre_treated, type = "pre-treated")
#> )
#> 
#> $plot_unsupervised_RF
#> ## Plot multidimensional scaling plot of unsupervised random forest
#> tar_target(
#>   plot_unsupervised_RF,
#>   metabolyseR::plotUnsupervisedRF(pre_treated, type = "pre-treated")
#> )
#> 
#> $plot_supervised_RF
#> ## Plot multidimensional scaling plot of supervised random forest
#> tar_target(
#>   plot_supervised_RF,
#>   metabolyseR::plotSupervisedRF(pre_treated, type = "pre-treated")
#> )
#> 

targetsMFassignment('FIE-HRMS fingerprinting')

#> $molecular_formula_assignment_parameters
#> ## Generate molecular formula assignment parameters
#> tar_target(
#>   molecular_formula_assignment_parameters,
#>   MFassign::assignmentParameters("FIE")
#> )
#> 
#> $molecular_formula_assignment
#> ## Perform molecular formula assignment
#> tar_target(
#>   molecular_formula_assignment,
#>   pre_treated %>% metabolyseR::dat(type = "pre-treated") %>% MFassign::assignMFs(molecular_formula_assignment_parameters),
#>   memory = "transient"
#> )
#> 
#> $assigned_data
#> ## Retieve pre-treated data with molecular formula assignments added to the feature names
#> tar_target(
#>   assigned_data,
#>   metaboMisc::addAssignments(pre_treated, molecular_formula_assignment)
#> )
#> 
#> $summary_assignments
#> ## Summarise the assigned molecular formulas
#> tar_target(
#>   summary_assignments,
#>   MFassign::summariseAssignment(molecular_formula_assignment)
#> )
#> 
#> $export_assignments
#> ## Export molecular formula assignments
#> tar_files(
#>   export_assignments,
#>   metaboMisc::export(molecular_formula_assignment, outPath = "exports/molecular_formula_assignments")
#> )
#> 

targetsModelling('FIE-HRMS fingerprinting')

#> $modelling_parameters
#> ## Detect appropriate modelling parameters
#> tar_target(
#>   modelling_parameters,
#>   metaboMisc::detectModellingParameters(assigned_data, cls = "class")
#> )
#> 
#> $modelling
#> ## Perform modelling
#> tar_target(
#>   modelling,
#>   metabolyseR::reAnalyse(assigned_data, modelling_parameters)
#> )
#> 
#> $plot_explanatory_heatmap
#> ## Plot a heat map of explanatory features
#> tar_target(
#>   plot_explanatory_heatmap,
#>   metabolyseR::plotExplanatoryHeatmap(modelling)
#> )
#> 
#> $summary_modelling_metrics
#> ## Retrieve modelling metrics
#> tar_target(
#>   summary_model_metrics,
#>   metabolyseR::metrics(modelling)
#> )
#> 
#> $summary_modelling_importance
#> ## Retireve modelling feature importance
#> tar_target(
#>   summary_model_importance,
#>   metabolyseR::importance(modelling)
#> )
#> 
#> $export_modelling
#> ## Export modelling results
#> tar_files(
#>   export_modelling,
#>   metaboMisc::exportModelling(modelling, outPath = "exports/modelling")
#> )
#> 

targetsCorrelations('FIE-HRMS fingerprinting')

#> $correlations_parameters
#> ## Generate parameters for correlation analysis
#> tar_target(
#>   correlations_parameters,
#>   metabolyseR::analysisParameters("correlations")
#> )
#> 
#> $correlations
#> ## Perform correlation analysis
#> tar_target(
#>   correlations,
#>   metabolyseR::reAnalyse(assigned_data, correlations_parameters)
#> )
#> 
#> $summary_correlations
#> ## Retrieve correlation analysis results
#> tar_target(
#>   summary_correlations,
#>   metabolyseR::analysisResults(correlations, "correlations")
#> )
#> 
#> $export_correlations
#> ## Export correlation analysis results
#> tar_files(
#>   export_correlations,
#>   metaboMisc::exportCorrelations(correlations, outPath = "exports/correlations")
#> )
#>